% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Calculation of the area under the ROC curve}
\usage{
auc(val, y)
}
\arguments{
\item{val}{Values of estimated intensities of a habitat map, based on which the area under the ROC curve is calculated}

\item{y}{Class label taking False or True based on which the area under the ROC curve is calculated}
}
\description{
\code{auc} returns the value of the area under the ROC curve
}
\seealso{
\code{\link{qPPP}}
}
