% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot0.R
\name{boxplot0}
\alias{boxplot0}
\title{Estimation of coefficients for Maxent, GM, rGM and Fisher}
\usage{
boxplot0(env = Env2, sp = sp_pteridium_aquilinum, iter = 500)
}
\arguments{
\item{sp}{index vector for presence locations.}

\item{iter}{maximum number of maximum iterations.}

\item{Xb}{data matrix of environmental variables and bias variables.}
}
\value{
estimated coefficients \eqn{\lambda} for Maxent, GM, rGM and Fisher
}
\description{
\code{boxplot0} returns the estimated coefficients \eqn{\lambda} for Maxent, GM, rGM and Fisher
}
\examples{
# Example usage
boxplot0()
}
