% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBA-package.R
\docType{package}
\name{CBA-package}
\alias{CBA-package}
\alias{CBA}
\title{Cumulant-based approximation for species distribution}
\description{
The CBA package provides methods for efficient and fast estimation
of species distribution using cumulant-based approximations.
}
\details{
CBA: Cumulant-based approximation for species distribution

This package, CBA, provides tools for efficient and fast estimation
for species distribution.
}
\examples{
library(CBA)
help(CBA)
}
\seealso{
\code{\link{maxent}}, \code{\link{rgm}}, \code{\link{gm}},
\code{\link{fisher0}}, \code{\link{habitat}},
\code{\link{nceas}}, \code{\link{est.nceas}},
\code{\link{sim}}, \code{\link{est.sim}}
}
\author{
Osamu Komori, Seikei University, Japan \email{komori@st.seikei.ac.jp}
}
\keyword{package}
