% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.nceas.R
\name{est.nceas}
\alias{est.nceas}
\title{Generate the estimation results about NCEAS data analysis by Maxent, GM, rGM and Fisher}
\usage{
est.nceas(
  Xb = nsw14.tr,
  Xb.pa = nsw14.te,
  gamma.default = -10^{
     -5
 },
  gamma.default2 = -0.5
)
}
\arguments{
\item{Xb}{training data matrix}

\item{Xb.te}{test data matrix}
}
\value{
boxplots of test AUC and computational costs by Maxent, GM, rGM and Fisher
}
\description{
\code{est.nceas} returns the results of NCEAS data analysis by Maxent, GM, rGM and Fisher
}
