% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.sim.R
\name{est.sim}
\alias{est.sim}
\title{Generate the estimation results about simulation analysis by Maxent, GM, rGM and Fisher by the Broyden-Fletcher-Goldfarb-Shanno (BFGS) algorithm}
\usage{
est.sim(
  rho = 0.5,
  n.grid = 100,
  m = 500,
  p = 50,
  alpha0 = seq(0, 0.1, length.out = p),
  type = "poisson"
)
}
\arguments{
\item{rho}{a parameter in correlation matrix}

\item{m}{presence sample size}

\item{p}{sample size}

\item{alpha0}{parameter to be estimated in log-linear model}
}
\value{
the values of test AUC and computational costs by Maxent, GM, rGM and Fisher
}
\description{
\code{est.sim} returns the results of simulation analysis by Maxent, GM, rGM and Fisher
}
