% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher0.R
\name{fisher0}
\alias{fisher0}
\title{Estimation function of Fisher}
\usage{
fisher0(
  env = Env2,
  sp = sp_pteridium_aquilinum,
  iter = 500,
  tau = "default",
  mul = 0.1
)
}
\arguments{
\item{sp}{index vector for presence locations. The default is the index vector for Pteridium aquilinum.}

\item{iter}{maximum number of iterations.}

\item{tau}{penalty term. A numeric value or "default" indicating that used for the linear feature in Maxent.}

\item{Xb}{data matrix of environmental variables. The default is the data matrix for Pteridium aquilinum.}

\item{gamma}{\eqn{\gamma} value that must be negative.}
}
\value{
estimated coefficients \eqn{\lambda} and number of iteration
}
\description{
\code{fisher0} returns the estimated coefficients and a iteration number.
}
\examples{
# Example usage
fisher0()
}
\references{
Phillips, S. J., & Dudík, M. (2008). Modeling of species distributions with Maxent: new extensions and a comprehensive evaluation. Ecography, 31, 161-175.
}
\seealso{
\code{\link{maxent}}, \code{\link{rgm}}, \code{\link{gm}}, \code{\link{gamma0}}
}
