% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma0.R
\name{gamma0}
\alias{gamma0}
\title{Estimation function of Gamma}
\usage{
gamma0(
  env = Env2,
  sp = sp_pteridium_aquilinum,
  iter = 500,
  gamma = -0.001,
  tau = "default",
  mul = 1
)
}
\arguments{
\item{sp}{index vector for presence locations. The default is the index vector for Pteridium aquilinum.}

\item{iter}{maximum number of iterations.}

\item{gamma}{gamma value in the loss function that must be negative.}

\item{tau}{penalty term. A numeric value or "default" indicating that used for the linear feature in Maxent.}

\item{mul}{multiplier applied to tau when "default" is used.}

\item{Xb}{data matrix of environmental variables. The default is the data matrix for Pteridium aquilinum.}
}
\value{
estimated coefficients and the number of iterations.
}
\description{
\code{gamma0} returns the estimated coefficients and a iteration number.
}
\examples{
# Example usage
gamma0()
}
\references{
Phillips, S. J., & DudíLLk, M. (2008). Modeling of species distributions with Maxent: new extensions and a comprehensive evaluation. Ecography, 31, 161-175.
}
\seealso{
\code{\link{maxent}}, \code{\link{rgm}}, \code{\link{gm}}, \code{\link{fisher0}}
}
