% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{habitat}
\alias{habitat}
\title{Generation of estimated habitat maps}
\usage{
habitat(env = Env2, sp = sp_pteridium_aquilinum, iter = 500)
}
\arguments{
\item{sp}{index vector for presence locations.}

\item{iter}{maximum number of maximum iterations.}

\item{Xb}{data matrix of environmental variables and bias variables.}
}
\value{
estimated habitat maps
}
\description{
\code{habitat} returns the estimated habitat maps by \code{\link{maxent}}, \code{\link{rgm}}, \code{\link{gm}} and \code{\link{fisher0}}.
}
