% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nceas.R
\name{nceas}
\alias{nceas}
\title{NCEAS data analysis by Maxent, GM, rGM and Fisher}
\usage{
nceas(region = "AWT")
}
\arguments{
\item{region}{name of region in NCEAS data ("AWT","CAN","NSW","NZ","SA","SWI"). The default is "AWT"}
}
\value{
boxplots of test AUC and computational costs by Maxent, GM, rGM and Fisher
}
\description{
\code{nceas} returns the results of NCEAS data analysis by Maxent, GM, rGM and Fisher
}
\examples{
# Example usage
nceas()
}
