% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Generate boxplots about simulation analysis by Maxent, GM, rGM and Fisher}
\usage{
sim(
  n.iter = 100,
  rho = 0.5,
  p = 50,
  alpha0 = seq(0, 0.1, length.out = p),
  type = "poisson"
)
}
\arguments{
\item{n.iter}{iteration number of simulation}

\item{rho}{correlation value}

\item{p}{number of variables}

\item{alpha0}{a vector indicating the true values of parameters to be estimated}

\item{type}{type of random variables}
}
\value{
boxplots of mean square erros and computational costs by Maxent, GM, rGM and Fisher
}
\description{
\code{sim} returns boxplots of simulation analysis by Maxent, GM, rGM and Fisher
}
\examples{
# Example usage
sim()
}
