\name{CBA-package}
\alias{CBA-package}
\alias{CBA}
\docType{package}
\title{Cumulant-based approximation for species distribution}
\description{
  This package, CBA, provides tools for efficient and fast estimation for species distribution.
}
\author{
  Osamu Komori, Seikei University, Japan \email{komori@st.seikei.ac.jp}
}
\keyword{package}
\seealso{
  \code{\link{maxent}}, \code{\link{rgm}}, \code{\link{gm}}, \code{\link{gamma0}}, \code{\link{fisher0}}, \code{\link{habitat}}, \code{\link{boxplot0}}, \code{\link{make.path}}, \code{\link{nceas}}, \code{\link{est.nceas}}, \code{\link{sim}}, \code{\link{est.sim}}
}
\examples{
  library(CBA)
  help(CBA)
}
