% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{gm}
\alias{gm}
\title{Estimation function of GM}
\usage{
gm(
  env = Env2,
  sp = sp_pteridium_aquilinum,
  iter = 500,
  tau = "default",
  mul = 0.1,
  gamma = -1
)
}
\arguments{
\item{sp}{index vector for presence locations. The default is the index vector for Pteridium aquilinum.}

\item{iter}{maximum number of iterations.}

\item{tau}{penalty term. A numeric value or "default" indicating that used for the linear feature in Maxent.}

\item{mul}{multiplier applied to tau when "default" is used.}

\item{Xb}{data matrix of environmental variables. The default is the data matrix for Pteridium aquilinum.}
}
\value{
estimated coefficients and the number of iterations.
}
\description{
\code{gm} returns the estimated coefficients and a iteration number.
}
\examples{
# Example usage
gm()
}
\references{
Phillips, S. J., & DudíLLk, M. (2008). Modeling of species distributions with Maxent: new extensions and a comprehensive evaluation. Ecography, 31, 161-175.
}
\seealso{
\code{\link{maxent}}, \code{\link{rgm}}, \code{\link{gamma0}}, \code{\link{fisher0}}
}
